+++ CHANGELOG.md
@@ -1,8 +1,5 @@
 # Changelog
 
-## v1.x.y
-* Add `gcp_exocompute` Terraform  module.
-
 ## v1.0.0
 * Add `gcp_project` Terraform module.
 * Add `gcp_archival_location` Terraform module.
@@ -10,3 +7,4 @@
 * Rename `aws_cnp_account` to `aws_iam_account` and move it to the `modules/` directory.
 * Update the `aws_exocompute` tests to use the `aws_iam_account` module.
 * Add support for the `KUBERNETES_PROTECTION` feature to the `aws_iam_account` module.
+* Add `gcp_exocompute` Terraform  module.